+++ pakyow-ui/CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * [added] Log errors encountered when setting up ui subscriptions
   * [fixed] Issue causing data subscriptions to never be expired for a web socket connection
+  * [added] WebSocket message handler for logging from the ui
 
 # 1.0